refactor: 패스워드 유틸 클래스 추가 - 도메인 별 패키지 분리 및 위치 변경
